# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventario de Redes",
        "description": "Control de creación, actualización, visualización y eliminación de registros de usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/auth/roles": {
            "get": {
                "tags": [
                    "Auth_users"
                ],
                "summary": "( Muestra todos los roles registrados)",
                "operationId": "1b04ebc6a5a34e070ce93cd0075969d8",
                "responses": {
                    "200": {
                        "description": "Datos recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos recuperados exitosamente"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "anyOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Role1Schema"
                                                    },
                                                    {
                                                        "$ref": "#/components/schemas/Role2Schema"
                                                    },
                                                    {
                                                        "$ref": "#/components/schemas/Role3Schema"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Role1Schema": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    }
                },
                "type": "object"
            },
            "Role2Schema": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "2"
                    },
                    "name": {
                        "type": "string",
                        "example": "Supervisor"
                    }
                },
                "type": "object"
            },
            "Role3Schema": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "3"
                    },
                    "name": {
                        "type": "string",
                        "example": "Usuario"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}